[#32] Moved configuration related code to separate header/source
